backend/requirements.txt
fastapi[standard]
uvicorn[standard]
smartapi-python
pandas
numpy
pandas-ta
pyotp
python-dotenv
sqlalchemy
aiosqlite
loguru
apscheduler
httpx
websockets
pydantic-settings
